Update : posts, project 공통 로직을 메소드로 분리
