+++ src/lib/directoryToHtml.ts
@@ -1,0 +1,34 @@
+import fs from 'fs'
+import path from 'path'
+import matter from 'gray-matter'
+
+import { unified } from 'unified'
+import remarkParse from 'remark-parse'
+import remarkUnwrapImages from 'remark-unwrap-images'
+import remarkRehype from 'remark-rehype'
+import rehypeExternalLinks from 'rehype-external-links'
+import rehypeStringify from 'rehype-stringify'
+
+const directoryToHtml = async (directory: string, id: string) => {
+  const fullPath = path.join(directory, `${id}.md`)
+  const fileContents = fs.readFileSync(fullPath, 'utf8')
+  const matterResult = matter(fileContents)
+
+  const processedContent = await unified()
+    .use(remarkParse)
+    .use(remarkUnwrapImages)
+    .use(remarkRehype)
+    .use(rehypeExternalLinks, { target: '_blank', rel: ['noopener'] })
+    .use(rehypeStringify)
+    .process(matterResult.content)
+
+  const contentHtml = processedContent.toString()
+
+  return {
+    id,
+    contentHtml,
+    ...matterResult.data
+  }
+}
+
+export default directoryToHtml
