+++ src/lib/posts.ts
@@ -2,12 +2,7 @@ import fs from 'fs'
 import path from 'path'
 import matter from 'gray-matter'
 
-import { unified } from 'unified'
-import remarkParse from 'remark-parse'
-import remarkUnwrapImages from 'remark-unwrap-images'
-import remarkRehype from 'remark-rehype'
-import rehypeExternalLinks from 'rehype-external-links'
-import rehypeStringify from 'rehype-stringify'
+import directoryToHtml from './directoryToHtml'
 
 const postsDirectory = path.join(process.cwd(), 'src/posts')
 
@@ -47,23 +42,5 @@ export function getAllPostIds() {
 }
 
 export async function getPostData(id : string) {
-  const fullPath = path.join(postsDirectory, `${id}.md`)
-  const fileContents = fs.readFileSync(fullPath, 'utf8')
-  const matterResult = matter(fileContents)
-
-  const processedContent = await unified()
-    .use(remarkParse)
-    .use(remarkUnwrapImages)
-    .use(remarkRehype)
-    .use(rehypeExternalLinks, { target: '_blank', rel: ['noopener'] })
-    .use(rehypeStringify)
-    .process(matterResult.content)
-
-  const contentHtml = processedContent.toString()
-
-  return {
-    id,
-    contentHtml,
-    ...matterResult.data
-  }
+  return directoryToHtml(postsDirectory, id)
 }
